Introduce GameSessionMode
A "game session" is a sequence of levels. It starts in-game with the first
level. Once a level is finished, the bonus-screen is shown, and the next
level in the sequence is started.

Currently, a game session can only be started with an empty player model,
starting at the specified level/episode and difficulty. Once save game
loading is implemented, it will also be possible to create a game session
with a loaded player model.

Once boss levels are implemented, this mode will also be responsible for
showing the "episode finished" story screens, as well as the high-score
entry/display, and going back to the main menu afterwards.