Adjust health value range for model
The HUD shows only 8 slices of health, and has a special animation for
"zero health", which is the state where Duke will die if he takes one
more hit. Previously, the model had a corresponding value range for health,
with 0 being "can take one more hit". But it makes more sense for the model
to define 0 as being dead, and adapt the HUD display accordingly so that
it shows the "can take one more hit" animation at 1 health.