Fix and simplify "low health" animation timing
The previous code is a holdover from when the overall game logic timing
was off, so the animation had to use this 1.5 factor in order to look
correct (that state is from before the project was published on github).
This is not necessary anymore now, and actually made the animation too
fast.