Change window creation mode on Mac OS
The advantage of FULLSCREEN over FULLSCREEN_DESKTOP is that game inputs
are not interpreted as system shortcuts. Otherwise, Ctrl+Left would be "switch
to left space" and switch away from the game window, for example. The
user would need to disable those shortcuts on a system-wide level, or
change the game's control scheme to avoid clashes with OS shortcuts.

Using non-desktop fullscreen mode means there is a slightly longer
switch-over when entering or exiting the game, but all the OS shortcut
clashes are avoided. This seems like a good trade-off.