Implement blue guard enemy
A mysterious humanoid in a blue space-suit. Seems to perform guard
duties as well as technical work for the Rigelatin. Most often found in
or near Rigelatin buildings/installations.

Quite different from most of the other enemies, which are usually robotic or
alien in nature, these could very well be human - or are they some kind
of android?