Only make out-of-bounds solid on the sides (left/right)
Don't block off the top/bottom of the map - if an entity moves past that,
it should be allowed to go out of the map. This is how it works in the
original game, and is required to implement the player falling out of the
map. It also makes it easier to implement projectile wall impact effects
correctly: Impact effects should not appear if a projectile leaves the map.
If we know that the map is only solid on the sides, it's sufficient to
check if a projectile's x position is out of bounds before showing the
impact effect.