Fully integrate BonusScreen
The bonus screen UI was alredy in the game before, but it was always
showing the "No bonus" text. Now, we get the set of achieved bonuses
from the GameRunner and pass it on to the BonusScreen. The score is also
adjusted accordingly now.

For now, GameRunner::achievedBonuses() returns dummy data, but this will
change in a later commit.