Fix assertion when dieing again after respawning at checkpoint
The death animation assigns a DestructionEffects component to the player's
entity. When the player respawns at a checkpoint, the entity is reused,
so it still has that DestructionEffects. When the player dies again, this
triggers an assertion in entity::assign as soon as we try to assign
DestructionEffects again.

To prevent this, remove the DestructionEffects component when respawning.