Change filename for user profile file
We keep looking for a file with the previous name in case we don't have
one with the new name, but we always save to the new name.

This prevents older versions of RigelEngine (which don't have Options as
part of the profile) from erasing our settings, since they will still
write to the old name.