Refactor: Extract function
Note that we are calling the extracted function from a different place
than where we extract the code from. This is intentional, as it makes
more sense for this to happen separately from creating the window - it's
theoretically possible to destroy and re-create the window multiple times,
but we only need to set these attributes once.