Fix JSON merging
The JSON merge algorithm wasn't correct, as it would cause values in
the base to become null when they weren't present in the extension.

This change fixes the algorithm to actually do what we want, adds a couple
of unit tests to ensure correct functionality, and adds some error
checking and explanatory comments.