Make DukeScriptRunner render into its own render target
This is a necessary step in order to implement window resizing.
Currently, DukeScriptRunner relies on the fact that the top-level
render-target is not automatically cleared every frame. But when resizing
the window, this causes problems, since the previously rendered image
will persist after a window size change, resulting in incorrect scaling
(since the image was rendered for a different window size).