Make main render target non-persistent
Previously, the top-level render target was not automatically cleared
on each frame, resulting in the last rendered image to persist. This
is a problem for implementing a resizable window, as the persistent
image would stay on screen at the wrong scale after resizing.

This commit changes this so that the top-level render target is cleared
on every frame. Other code is adapted accordingly, most notably, we now
have to ensure that we rendered something before triggering a fade-out.