Fix CI build on Windows
std::result_of is deprecated in C++ 17, std::invoke_result should be used
instead. Visual Studio has a compiler warning for this, but so far, we
didn't run into it. But AppVeyor has recently updated its Visual Studio version
from 19.24.28316.0 to 19.25.28610.4, and the latter is now producing a
warning, making the build fail on CI.

To fix this, switch to using invoke_result.