Implement correct collision detection for player projectiles
This fixes a problem on M5, where Duke's rocket flies through a wooden
wall instead of setting it on fire in case Duke is standing right next
to the wall. Most likely, there are other places where this problem occurs.

There are two aspects to the fix:

* doing collision detection the same way as the original executable (see
  comments in code)
* doing collision checking on a projectile that has just been spawned and
  is thus not moving yet

The latter was a problem because we were relying on the PhysicsSystem
setting the CollidedWithWorld tag, but this only happens if the projectile
is already in motion. Now, we do the collision detection ourselves instead
of using the PhysicsSystem, which means we can do the check even for a
projectile that's not moving yet.