Implement top-level ingame menu
In the original game, the various ingame menus are only reachable via
direct keyboard shortcuts, like pressing F2 to open the "save game"
menu. This does not translate well to a gamepad, which does not have
that many buttons to spare.

In order to make the ingame menus usable with a gamepad, we therefore
introduce a completely new menu to act as a top-level entry point from
which the other ingame menus can be reached. This menu can be opened by
pressing the gamepad's "start" button (located to the right of the Xbox
logo button on an Xbox controller, to the right of the touchpad on a PS4
controller etc.).

At the moment, this new menu does not allow access to the options menu,
since the latter is not yet usable with a gamepad.