Take actual sample rate and audio format into account
Previously, the sound code assumed that the audio device would always
use the sample rate and audio format that were requested when calling
Mix_OpenAudio. While this is true for a lot of environments, it is not
guaranteed. Running inside a browser via WebAssembly is an example where
the audio format happens to be different from signed 16-bit on all the
systems I've tested.

With this commit, the code now queries SDL for the actual audio format
and sample rate used by the audio device, and performs any necessary
conversion as needed.