Cleanup: Remove ineffective culling code
Since the introduction of widescreen mode, the isVisible function stopped
working correctly, as it doesn't take the current transformation matrix into
account. It used to work before because all rendering was happening in the 320
x 200 coordinate system, but that's not the case anymore. Therefore, the
function would consider a lot of things visible that actually were outside of
the screen. This is not a problem on most hardware, which is why it went
unnoticed for so long.

Now that we do proper culling in the SpriteRenderingSystem, we can simply get
rid of the broken culling inside the renderer.