Fix mouse input on Raspberry Pi
In SDL versions before 2.0.14, SDL_GetGlobalMouseState returns 0,0 if the
current SDL backend does not support a global mouse position. This is the case
for the RPI backend, thus causing mouse interaction to be broken on the
Raspberry Pi. By disabling the use of SDL_GetGlobalMouseState for all SDL
backends that don't support it, we fix that problem.