Fix wide-screen switch on window resize
When running in windowed mode and the window is sized to have a non-wide
aspect ratio, wide-screen will not be enabled right away. Once the window
is resized to have a wide (> 4:3) aspect ratio, it will become active
automatically. Previously, there was a bug though causing part of the view
to be black. This fixes it.