Enhance the health bar for episode 3's boss
In the original game, the health bar is bigger than the viewport, making
it so that no change is visible in the health bar until a certain point,
at which it starts moving down quite quickly.

Instead, we show a proportional health bar for this boss, which reacts
immediately as the boss is damaged regardless of viewport size.