Fix sound crackling issues
Turns out we need to use a fresh AdLib emulator for each sound, otherwise
there are artifacts. Another solution would be to keep rendering more
samples to "drain" the AdLib, but that would be slower than simply
resetting the emulator.

Maybe worth investigating: Can we reset an existing emulator instead of
recreating it to be a little bit faster?