Emulate 16-bit far pointer arithmetic to recreate out-of-bounds reads
In 16-bit C code for DOS, far pointers are made up of two 16-bit values, a
segment and an offset. Both values are combined to form a linear address
when dereferencing the pointer. Crucially, incrementing/decrementing a far
pointer using C's regular language features only affects the offset part of
the pointer. This causes the pointer to wrap back around to a smaller
address in case the offset overflows the 16-bit value range (i.e
incremented beyond 0xFFFF).

This is relevant for making the collision detection code behave exactly
like it does in the original game. The collision detection is sometimes
invoked with out-of-range values, for example when an actor is positioned
near the left edge of the map (i.e. x 0 or close to 0), and then checks a
position like e.g. `x - 2`. Since the position is given as an unsigned
value, this can result in x being something like 65534.

The collision detection code doesn't have a lot of bounds checking, so in
many cases, an out-of-range value like that will be used in pointer
arithmetic expressions. In a linear memory model like we have in modern
32-bit or 64-bit code, this simply causes a read of memory beyond the map
data array. But in the original 16-bit code, the result is a bit more
interesting. The map data array itself is 65500 bytes in size. The far
pointer referencing it has a maximum valid range covering 65535 bytes.
Overflowing that pointer's offset by using an x value that's close to
65535 thus usually causes the resulting address to still be inside of the
map data, just at a somewhat arbitrary location.

In order to recreate how the original code behaves, we also need to
recreate this pointer arithmetic. We do that by using a 16-bit unsigned int
(`word`) instead of a pointer for all arithmetic, and then only actually
dereferencing the map data when we have the final value.

Notably, there are some edge cases where this scheme doesn't fully
recreate the original behavior. As mentioned above, the map data is only
65500 bytes, but the value range for a far pointer's offset is 65535.
It's therefore possible to actually overflow the map data array, even with
the offset's wrap-around behavior. If that happens, the memory that's read
will be whatever memory has been allocated after allocating the map data
using the game's memory manager. This will be map data copies in case the
map contains elevators or sliding door actors, otherwise it will be
whatever random old garbage data is found at that location.

This edge case seems unlikely to occur in practice, so we don't attempt
to handle it for now.